improved postinst script to be more verbose and allow for errors
--

it is not overly important that this killing of gpg-agents always
works. we did without it before, and therefore don't want to worry
about permission problems or errors because no agent was running
at all. also inform users that agents were killed and how one
can be started manually if you need one for ssh now.